fix: convert Float to DECIMAL(10,2) for monetary columns in eBay models
Replace Float (double precision) with DECIMAL(10,2) for exact monetary
precision in 10 columns across 6 eBay models (orders, returns, refunds,
payment disputes, inquiries, cancellations).

+++ backend/models/user/ebay_cancellation.py
@@ -12,9 +12,10 @@ Business Rules:
 """
 
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
-from sqlalchemy import BigInteger, DateTime, Float, ForeignKey, Text, func
+from sqlalchemy import BigInteger, DateTime, DECIMAL, ForeignKey, Text, func
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -116,7 +117,7 @@ class EbayCancellation(Base):
     )
 
     # Refund
-    refund_amount: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
+    refund_amount: Mapped[Optional[Decimal]] = mapped_column(DECIMAL(10, 2), nullable=True)
     refund_currency: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
     refund_status: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
