+++ backend/models/user/ebay_refund.py
@@ -12,9 +12,10 @@ Business Rules:
 """
 
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
-from sqlalchemy import BigInteger, DateTime, Float, ForeignKey, Text, func
+from sqlalchemy import BigInteger, DateTime, DECIMAL, ForeignKey, Text, func
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -94,9 +95,9 @@ class EbayRefund(Base):
     )  # PENDING, REFUNDED, FAILED
 
     # Amount
-    refund_amount: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
+    refund_amount: Mapped[Optional[Decimal]] = mapped_column(DECIMAL(10, 2), nullable=True)
     refund_currency: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    original_amount: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
+    original_amount: Mapped[Optional[Decimal]] = mapped_column(DECIMAL(10, 2), nullable=True)
 
     # Reason
     reason: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
